refactor(back,xpsys): move xp systems logic to xp-systems
