+++ libs/xp-systems/src/xp-system.class.spec.ts
@@ -1,21 +1,7 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { XpSystemsService } from './xp-systems.service';
-import { mockDeep } from 'jest-mock-extended';
-import { PRISMA_MOCK_PROVIDER } from '../../../../test/prisma-mock.const';
+import { XpSystems } from './xp-systems.class';
 
 describe('XpSystemsService', () => {
-  let service: XpSystemsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [XpSystemsService, PRISMA_MOCK_PROVIDER]
-    })
-      .useMocker(mockDeep)
-      .compile();
-
-    service = module.get<XpSystemsService>(XpSystemsService);
-    await service.onModuleInit();
-  });
+  const service = new XpSystems();
 
   it('should be defined', () => {
     expect(service).toBeDefined();
@@ -26,6 +12,7 @@ describe('XpSystemsService', () => {
       const xp = service.getCosmeticXpInLevel(0);
       expect(xp).toEqual(0);
     });
+
     it('should handle greater than 0', () => {
       const xp = service.getCosmeticXpInLevel(1);
       expect(xp).toBeGreaterThan(0);
@@ -37,6 +24,7 @@ describe('XpSystemsService', () => {
       const xp = service.getCosmeticXpForLevel(0);
       expect(xp).toEqual(0);
     });
+
     it('should handle greater than 0', () => {
       const xp = service.getCosmeticXpForLevel(1);
       expect(xp).toBeGreaterThan(0);
